requirements.txt
asgiref>=3.8.1,<4.0.0
certifi>=2024.6.2,<2025.0.0
cffi==1.16.0
charset-normalizer>=3.3.2,<3.4.0
cryptography>=42.0.7,<43.0.0
defusedxml==0.8.0rc2
Django==5.0.6
django-templated-mail>=1.1.1,<1.2.0
djangorestframework>=3.15.1,<3.16.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
idna>=3.7,<3.8.0
oauthlib>=3.2.2,<3.3.0
pycparser>=2.22,<2.23.0
PyJWT>=2.8.0,<2.9.0
python3-openid>=3.2.0,<3.3.0
pytz>=2024.1,<2025.0.0
requests>=2.32.3,<2.33.0
requests-oauthlib>=2.0.0,<2.1.0
social-auth-app-django>=5.4.1,<5.5.0
social-auth-core>=4.5.4,<4.6.0
sqlparse>=0.5.0,<0.6.0
urllib3>=2.2.1,<2.3.0
django-allauth
coreapi
drf_yasg
